Cannot find node issue when calculating new routes I have created a new list of station with truncated geom and are using this in the nearest node function. This appears to have stopped this error.
There is still a issue with networkx returning some routes which have a length of zero and  are technically valid. This causes an error in the visualisation, though I catch has been put in place which treates these as instances where no route was possible, be this the case or not.